+++ src/components/offer-list/offer-list.test.tsx
