+++ src/components/page-header/header-pic.tsx
