+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/App/app';
+import App from './components/app/app';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import { fetchOffers, userCheckAuth } from './store/actionsAPI';
