+++ src/pages/offer-page/offer-restructor.ts
