+++ src/store/reducers/city-reducer.ts
@@ -1,5 +1,5 @@
 import {createReducer} from '@reduxjs/toolkit';
-import { City } from '../../components/Map/map';
+import { City } from '../../components/map/map';
 import { CITIES } from '../../recources/cities';
 import { changeCity } from '../actions';
